Moved static functions to new file
